Extract P25 receiver core and add live validation
Move the hardware-independent Phase I C4FM protocol, FEC, trunking state,
and voice-frame extraction out of the Tab5 adapter. The core now accepts
explicit timestamps, uses fixed memory, and delivers voice frames through a
bounded callback. Move mbelib deinterleave and PCM synthesis into a separate
hardware-independent voice module while leaving FreeRTOS tasks, tuning policy,
audio queues, and speaker delivery in the Tab5 application.

Add authenticated P25 control-channel IQ capture and replay commands, a
session-level grant-event counter, generic over-the-air identity reporting,
and a repeatable validation runner. P25 capture is control-only, capped at
one MiB including the ORCIQ header, and clears its ready state after a
successful SD export so repeated captures do not require a reboot.

Add a one-MiB Lane County control-channel fixture with recorded provenance
and SHA-256, deterministic replay assertions, allocation-free stress checks,
optimized and ASan/LSan/UBSan host runs, and a narrow GitHub Actions job.
Update architecture, porting, serial CLI, dashboard, and feature-status docs.
Public status remains P25 WIP because wider-system and CQPSK/LSM compatibility
are later acceptance gates.

Validation performed:
- P25 host tests passed optimized and under ASan/LSan/UBSan.
- Native ESP-IDF 5.5.4 Tab5 build passed.
- Final firmware SHA-256 was 6E015C34DEECB31DD0F66A23C6AEB41FC0B000828A52F6DD866EB13B186F95F6.
- That exact image flashed to COM17 with esptool verification.
- Automated dashboard, transition, font-restoration, Wi-Fi, and RTL-SDR regression passed.
- Live Lane County validation locked 453.925 MHz, followed five grant events,
  produced 846 IMBE frames and 812,160 PCM samples, returned to control, and
  relocked with zero USB, IQ, audio, and voice-queue drops.
- The saved fixture replayed to NAC 1F0, WACN BEE00, system 1F3, RFSS/site 1/1,
  seven valid NIDs, and twenty valid TSBKs with no voice frames.
- The project owner accepted the P25 display and confirmed intelligible clear
  voice on multiple real talkgroups through the 3.5 mm output. Indoor signal
  quality still varied by voice channel and antenna placement.
