+++ .github/workflows/p25-core-tests.yml
@@ -1,0 +1,34 @@
+name: P25 core
+
+on:
+  pull_request:
+    paths:
+      - apps/orcsdr-tab5/ui/p25_decoder_core.*
+      - apps/orcsdr-tab5/ui/p25_voice.*
+      - components/mbelib/**
+      - tests/fixtures/p25_control_lane_453925.orciq
+      - tests/p25_core_tests.cpp
+      - tools/test-p25-core.*
+      - .github/workflows/p25-core-tests.yml
+  push:
+    branches: [main]
+    paths:
+      - apps/orcsdr-tab5/ui/p25_decoder_core.*
+      - apps/orcsdr-tab5/ui/p25_voice.*
+      - components/mbelib/**
+      - tests/fixtures/p25_control_lane_453925.orciq
+      - tests/p25_core_tests.cpp
+      - tools/test-p25-core.*
+      - .github/workflows/p25-core-tests.yml
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@11d5960a326750d5838078e36cf38b85af677262 # v4
+      - name: Optimized and sanitizer tests
+        run: bash tools/test-p25-core.sh
