+++ apps/orcsdr-tab5/ui/p25_decoder.hpp
@@ -1,64 +1,23 @@
 #pragma once
 
-#include <cstddef>
-#include <cstdint>
+#include "p25_decoder_core.hpp"
 
 namespace orcsdr::p25decoder {
 
-constexpr size_t kRecentGrantCount = 8;
-constexpr size_t kVoiceFrameBits = 144;
+using p25core::Grant;
+using p25core::Snapshot;
+using p25core::VoiceFrame;
+inline constexpr size_t kRecentGrantCount = p25core::kRecentGrantCount;
+inline constexpr size_t kVoiceFrameBits = p25core::kVoiceFrameBits;
 
-struct VoiceFrame {
-  uint8_t bits[kVoiceFrameBits]{};
-  uint32_t sequence = 0;
-};
-
-struct Grant {
-  bool valid = false;
-  bool encrypted = false;
-  bool emergency = false;
-  bool tdma = false;
-  uint16_t talkgroup = 0;
-  uint32_t source_id = 0;
-  uint32_t frequency_hz = 0;
-  uint32_t seen_ms = 0;
-};
-
-struct Snapshot {
-  bool frame_sync = false;
-  bool identity_valid = false;
-  uint16_t nac = 0;
-  uint32_t wacn = 0;
-  uint16_t system_id = 0;
-  uint8_t rfss = 0;
-  uint8_t site = 0;
-  uint32_t sync_words = 0;
-  uint32_t nid_good = 0;
-  uint32_t nid_failed = 0;
-  uint32_t nid_corrected_bits = 0;
-  uint32_t tsbk_good = 0;
-  uint32_t tsbk_failed = 0;
-  uint32_t voice_ldus = 0;
-  uint32_t voice_frames = 0;
-  uint32_t voice_queue_drops = 0;
-  uint32_t last_voice_ms = 0;
-  uint16_t last_trellis_metric = 0;
-  float estimated_ber_percent = 0.0f;
-  float frame_error_percent = 0.0f;
-  float afc_offset_hz = 0.0f;
-  float symbol_level = 0.0f;
-  Grant current_grant{};
-  Grant recent_grants[kRecentGrantCount]{};
-};
-
-// The decoder is single-writer: call reset/process_cu8 from the RTL delivery
-// task. snapshot() is safe from the UI task.
+// Device adapter: the RTL delivery task is the single decoder writer. The UI
+// snapshot and voice-task queue are synchronized here, outside the core.
 void reset();
 void process_cu8(const uint8_t* iq, size_t bytes);
+void reset_at(uint32_t now_ms);
+void process_cu8_at(const uint8_t* iq, size_t bytes, uint32_t now_ms);
 Snapshot snapshot();
 bool pop_voice_frame(VoiceFrame* frame);
-
-// Deterministic protocol/FEC check. It does not touch live decoder state.
 bool self_check();
 
 }  // namespace orcsdr::p25decoder
