+++ docs/user-guide/dashboards/p25.md
@@ -1,6 +1,8 @@
 # P25
 
-P25 is experimental receive-only monitoring. OrcSDR follows decoded clear voice grants and skips encryption when configured.
+**P25 WIP — Phase I clear voice verified; wider system compatibility in progress.**
+
+P25 is receive-only monitoring. OrcSDR follows decoded clear voice grants and skips encryption when configured.
 
 ## Monitor workflow
