+++ docs/user-guide/feature-status.md
@@ -4,7 +4,7 @@
 |---|---|---|
 | FM receive, stereo audio, presets, and health | Implemented | Live hardware verified |
 | FM RDS decoding | Implemented | Sequential A–B–C–D groups, confirmed PS, voted PTY, Radio Text. Needs a 260 kHz FM filter. This dongle uses a +13 kHz LO bias that is not shown as the channel. |
-| P25 control and clear voice following | Experimental | Encrypted voice is not decoded |
+| P25 control and clear voice following | WIP | Phase I clear voice verified; wider system compatibility in progress. Encrypted voice is not decoded. |
 | ADS-B 1090 dashboard and aircraft database | Implemented | Live hardware verified; coverage depends on antenna, location, and valid position messages |
 | LoRa receive, packet views, and capture | Experimental | No transmit path |
 | AM, WX, CB, and Browse tools | Experimental | Shared radio/scope/capture foundation |
