+++ tests/fixtures/README.md
@@ -1,0 +1,12 @@
+# P25 replay fixture
+
+`p25_control_lane_453925.orciq` is a 1,048,576-byte ORCIQ capture made by the
+project owner on 2026-09-05 from the Lane County, Oregon P25 Phase I control
+channel at 453.925 MHz. It contains 1,048,540 bytes of unsigned 8-bit
+interleaved I/Q sampled at 960 kS/s. Voice following was disabled during the
+capture, so the fixture contains no intentionally captured voice traffic.
+
+SHA-256: `33d884a99dd7cf0332b02df0f4bad57d54d92e2587db83760e951688440b7932`
+
+The deterministic receiver result is NAC `1F0`, WACN `BEE00`, system `1F3`,
+RFSS/site `1/1`, 7 valid NIDs, and 20 valid TSBKs.
